Reminder Creation
showing what I have completed so far. Before I go into AWS to create an API for the reminder creation, I wanted to run the code by you to see if you have it.

On my caregiver code, its just missing the API address. Let me know if there is anything else I need to fix on this code or if its off. I have this code worked towards storing the data into dynamodb. If this is not correct, let me know.